+++ backend/src/baserow/contrib/database/views/exceptions.py
@@ -8,6 +8,18 @@ class ViewDoesNotExist(Exception):
     """Raised when trying to get a view that doesn't exist."""
 
 
+class ViewNotInTable(Exception):
+    """Raised when a provided view does not belong to a table."""
+
+    def __init__(self, view_id=None, *args, **kwargs):
+        self.view_id = view_id
+        super().__init__(
+            f"The view {view_id} does not belong to the table.",
+            *args,
+            **kwargs,
+        )
+
+
 class UnrelatedFieldError(Exception):
     """
     Raised when a field is not related to the view. For example when someone tries to