+++ backend/src/baserow/contrib/database/views/signals.py
@@ -4,6 +4,7 @@ from django.dispatch import Signal
 view_created = Signal()
 view_updated = Signal()
 view_deleted = Signal()
+views_reordered = Signal()
 
 view_filter_created = Signal()
 view_filter_updated = Signal()