+++ backend/src/baserow/contrib/database/ws/signals.py
@@ -1,5 +1,5 @@
 from .table.signals import table_created, table_updated, table_deleted
-from .views.signals import view_created, view_updated, view_deleted
+from .views.signals import view_created, views_reordered, view_updated, view_deleted
 from .rows.signals import row_created, row_updated, row_deleted
 from .fields.signals import field_created, field_updated, field_deleted
 
@@ -8,6 +8,7 @@ __all__ = [
     "table_created",
     "table_updated",
     "table_deleted",
+    "views_reordered",
     "view_created",
     "view_updated",
     "view_deleted",