+++ web-frontend/modules/database/realtime.js
@@ -149,6 +149,13 @@ export const registerRealtimeEvents = (realtime) => {
     }
   })
 
+  realtime.registerEvent('views_reordered', ({ store, app }, data) => {
+    const table = store.getters['table/getSelected']
+    if (table !== undefined && table.id === data.table_id) {
+      store.commit('view/ORDER_ITEMS', data.order)
+    }
+  })
+
   realtime.registerEvent('view_deleted', ({ store }, data) => {
     const view = store.getters['view/get'](data.view_id)
     if (view !== undefined) {